visualizer/requirements.txt
aiofiles==22.1.0
async-generator==1.10
attrs==22.1.0
blinker==1.5
CacheControl==0.12.11
cachelib==0.9.0
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
chromedriver-binary==108.0.5359.71.0
click==8.1.3
cryptography==38.0.4
dash==2.7.0
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-extensions==0.1.8
dash-html-components==2.0.0
dash-interactive-graphviz==0.3.0
dash-renderjson==0.0.1
dash-table==5.0.0
dill==0.3.6
diskcache==5.4.0
EditorConfig==0.12.3
exceptiongroup==1.0.4
firebase-admin==6.0.1
Flask==2.2.2
Flask-Caching==2.0.1
google-api-core==2.11.0
google-api-python-client==2.69.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.3
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
grpcio-status==1.51.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httplib2==0.21.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsbeautifier==1.14.7
loguru==0.6.0
MarkupSafe==2.1.1
more-itertools==8.14.0
msgpack==1.0.4
multiprocess==0.70.14
numpy==1.23.5
outcome==1.2.0
pandas==1.5.2
Pillow==9.3.0
plotly==5.11.0
priority==2.0.0
proto-plus==1.22.1
protobuf==4.21.11
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
PyJWT==2.6.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
quart==0.18.3
requests==2.28.1
rsa==4.9
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tenacity==8.1.0
toml==0.10.2
trio==0.22.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.13
Werkzeug==2.2.2
wsproto==1.2.0
